Removed all TinyXML usage and references. Tweaked Indigo::Vector resizing a bit.
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@5921 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ utils/SystemInfo.cpp
@@ -137,7 +137,7 @@ void SystemInfo::getMACAddresses(std::vector<std::string>& addresses_out)
 	int num_calls_done = 0;
 	while(1) // Loop until GetAdaptersInfo() succeeds or we exceed the max number of attempts.
 	{
-		DWORD dwBufLen = AdapterInfo.size() * sizeof(IP_ADAPTER_INFO); // The size of the buffer.
+		DWORD dwBufLen = (DWORD)(AdapterInfo.size() * sizeof(IP_ADAPTER_INFO)); // The size of the buffer.
 		DWORD dwStatus = GetAdaptersInfo(
 			AdapterInfo.data(),	// [out] buffer to receive data
 			&dwBufLen		// [in] size of receive data buffer
@@ -146,14 +146,14 @@ void SystemInfo::getMACAddresses(std::vector<std::string>& addresses_out)
 
 		// If we get an ERROR_BUFFER_OVERFLOW it means the buffer was too small.
 		// Allocate a buffer that is large enough and call GetAdaptersInfo again.
-		if (dwStatus == ERROR_BUFFER_OVERFLOW)
+		if(dwStatus == ERROR_BUFFER_OVERFLOW)
 		{
 			AdapterInfo.resize(dwBufLen / sizeof(IP_ADAPTER_INFO));
 
 			dwStatus = GetAdaptersInfo(
 				AdapterInfo.data(),	// [out] buffer to receive data
 				&dwBufLen		// [in] size of receive data buffer
-				);
+			);
 		}
 
 		if(dwStatus == ERROR_SUCCESS)
